fix: migrate NextAuth route to App Router format with named exports
