+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,17 +1,5 @@
 import NextAuth from 'next-auth'
-import GitHubProvider from 'next-auth/providers/github'
-import type { NextAuthOptions } from 'next-auth'
-import { env } from '@/utils/env'
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GitHubProvider({
-      clientId: env.GITHUB_ID,
-      clientSecret: env.GITHUB_SECRET,
-    }),
-  ],
-  secret: env.NEXTAUTH_SECRET,
-}
+import { authOptions } from '@/lib/auth'
 
 const handler = NextAuth(authOptions)
